+++ lib/detectable_text_field.dart
@@ -1,6 +1,7 @@
 library detectable_text_field;
 
-export 'package:detectable_text_field/widgets/detectable_text_field_v2.dart';
-export 'package:detectable_text_field/detectable_text_field.dart';
+export 'package:detectable_text_field/widgets/detectable_text_field.dart';
+export 'package:detectable_text_field/widgets/detectable_text_editing_controller.dart';
+export 'package:detectable_text_field/widgets/detectable_text.dart';
 
-export 'widgets/detectable_text.dart';
+export 'package:detectable_text_field/detectable_text_field.dart';
