--- lib/widgets/detectable_text_field_v2.dart
@@ -1,84 +1,0 @@
-import 'package:detectable_text_field/widgets/detectable_text_editing_controller.dart';
-import 'package:flutter/material.dart';
-
-class DetectableTextFieldV2 extends TextField {
-  DetectableTextFieldV2({
-    super.key,
-    RegExp? regExp,
-    DetectableTextEditingController? controller,
-    TextStyle? detectedStyle,
-    super.focusNode,
-    super.undoController,
-    super.decoration,
-    super.keyboardType,
-    super.textInputAction,
-    super.textCapitalization,
-    super.style,
-    super.strutStyle,
-    super.textAlign,
-    super.textAlignVertical,
-    super.textDirection,
-    super.readOnly,
-    super.showCursor,
-    super.autofocus,
-    super.obscuringCharacter,
-    super.obscureText,
-    super.autocorrect,
-    super.smartDashesType,
-    super.smartQuotesType,
-    super.enableSuggestions,
-    super.maxLines,
-    super.minLines,
-    super.expands,
-    super.maxLength,
-    super.maxLengthEnforcement,
-    super.onChanged,
-    super.onEditingComplete,
-    super.onSubmitted,
-    super.onAppPrivateCommand,
-    super.inputFormatters,
-    super.enabled,
-    super.cursorWidth,
-    super.cursorHeight,
-    super.cursorRadius,
-    super.cursorOpacityAnimates,
-    super.cursorColor,
-    super.selectionHeightStyle,
-    super.selectionWidthStyle,
-    super.keyboardAppearance,
-    super.scrollPadding,
-    super.dragStartBehavior,
-    super.enableInteractiveSelection,
-    super.selectionControls,
-    super.onTap,
-    super.onTapOutside,
-    super.mouseCursor,
-    super.buildCounter,
-    super.scrollController,
-    super.scrollPhysics,
-    super.autofillHints,
-    super.contentInsertionConfiguration,
-    super.clipBehavior,
-    super.restorationId,
-    super.scribbleEnabled,
-    super.enableIMEPersonalizedLearning,
-    super.contextMenuBuilder,
-    super.canRequestFocus,
-    super.spellCheckConfiguration,
-    super.magnifierConfiguration,
-  })  : assert(
-          !(regExp != null && controller != null),
-          'Cannot provide both a regExp and a controller. Use a controller to define regExp.\n Example: DetectableTextEditingController(regExp: yourRexExp)',
-        ),
-        assert(
-          !(detectedStyle != null && controller != null),
-          'Cannot provide both a detectedStyle and a controller. Use a controller to define detectedStyle.\n Example: DetectableTextEditingController(detectedStyle : yourDetectedStyle)',
-        ),
-        super(
-          controller: controller ??
-              DetectableTextEditingController(
-                regExp: regExp,
-                detectedStyle: detectedStyle,
-              ),
-        );
-}
